+++ packages/polaris-test/nest-server-code-first/graphql/modules/pen.module.ts
@@ -1,4 +1,4 @@
-import { TypeOrmModule } from '@enigmatis/polaris-nest';
+import { PolarisTypeORMInjector, TypeOrmModule } from '@enigmatis/polaris-nest';
 import { Module } from '@nestjs/common';
 import { Author } from '../../../shared-resources/entities/author';
 import { Book } from '../../../shared-resources/entities/book';
@@ -8,6 +8,6 @@ import { Pen } from '../../../shared-resources/entities/pen';
 
 @Module({
     imports: [TypeOrmModule.forFeature([Pen, Book, Author])],
-    providers: [PenResolver, PenService],
+    providers: [PenResolver, PenService, PolarisTypeORMInjector],
 })
 export class PenModule {}