+++ packages/polaris-test/nest-server-schema-first/graphql/services/book.service.ts
@@ -3,16 +3,14 @@ import {
     Edge,
     Like,
     PageConnection,
-    PolarisConnection,
     PolarisRepository,
 } from '@enigmatis/polaris-core';
 import { Inject, Injectable, Scope } from '@nestjs/common';
 import { CONTEXT } from '@nestjs/graphql';
-import { InjectConnection } from '@nestjs/typeorm';
 import { PubSubEngine } from 'graphql-subscriptions';
 import { Author } from '../../../shared-resources/entities/author';
 import { Book } from '../../../shared-resources/entities/book';
-import { OnlinePagingInput } from '@enigmatis/polaris-nest';
+import { OnlinePagingInput, PolarisTypeORMInjector } from '@enigmatis/polaris-nest';
 import { TestContext } from '../../../shared-resources/context/test-context';
 
 const BOOK_UPDATED = 'BOOK_UPDATED';
@@ -22,13 +20,13 @@ export class BookService {
     private bookRepository: PolarisRepository<Book>;
     private authorRepository: PolarisRepository<Author>;
     constructor(
-        @InjectConnection()
-        connection: PolarisConnection,
         @Inject(CONTEXT) private readonly ctx: TestContext,
+        @Inject(PolarisTypeORMInjector)
+        private readonly polarisTypeORMInjector: PolarisTypeORMInjector,
         @Inject('PUB_SUB') private pubSub: PubSubEngine,
     ) {
-        this.bookRepository = connection.getRepository(Book, ctx);
-        this.authorRepository = connection.getRepository(Author, ctx);
+        this.bookRepository = this.polarisTypeORMInjector.getRepository(Book);
+        this.authorRepository = this.polarisTypeORMInjector.getRepository(Author);
     }
 
     public async findAll(): Promise<any[]> {